media: intel-ipu4: fix i386 build warnings
Message for Open Source:
Fixed warnings when build with i386 support:

make ARCH=i386

ipu-fw-com.c:280:24: warning: cast from pointer to integer of
different size_t [-Wpointer-to-int-cast]
ipu-buttress.c:868:37: warning: right shift count >= width of type
[-Wshift-count-overflow]
libintel-ipu4.c: warning: comparison between enum
ipu_fw_isys_resolution_info and enum ia_css_isys_resolution_info
[-Wenum-compare]
__udivdi3 [drivers/media/pci/intel/ipu4/intel-ipu4.ko] undefined!
...

Message for Internal:
[Issue/Feature] same as above.
[Root Cause/Changes] same as above.

Tracked-On: #JICCCN-228

+++ drivers/media/pci/intel/ipu-fw-com.c
@@ -273,11 +273,11 @@ void *ipu_fw_com_prepare(struct ipu_fw_com_cfg *cfg,
 	ctx->specific_vied_addr = ctx->dma_addr + offset;
 	offset += roundup(cfg->specific_size, 8);
 
-	ctx->ibuf_host_addr = (u64)(ctx->dma_buffer + offset);
+	ctx->ibuf_host_addr = (uintptr_t)(ctx->dma_buffer + offset);
 	ctx->ibuf_vied_addr = ctx->dma_addr + offset;
 	offset += sizeinput;
 
-	ctx->obuf_host_addr = (u64)(ctx->dma_buffer + offset);
+	ctx->obuf_host_addr = (uintptr_t)(ctx->dma_buffer + offset);
 	ctx->obuf_vied_addr = ctx->dma_addr + offset;
 	offset += sizeoutput;
 
@@ -415,7 +415,7 @@ void *ipu_send_get_token(struct ipu_fw_com_context *ctx, int q_nbr)
 
 	index = curr_index(q_dmem, DIR_SEND);
 
-	return (void *)q->host_address + (index * q->token_size);
+	return (void *)(unsigned long)q->host_address + (index * q->token_size);
 }
 EXPORT_SYMBOL_GPL(ipu_send_get_token);
 
@@ -424,7 +424,8 @@ void ipu_send_put_token(struct ipu_fw_com_context *ctx, int q_nbr)
 	struct sys_queue *q = &ctx->input_queue[q_nbr];
 	void __iomem *q_dmem = ctx->dmem_addr + q->wr_reg * 4;
 	int index = curr_index(q_dmem, DIR_SEND);
-	void *addr = (void *)q->host_address + (index * q->token_size);
+	void *addr = (void *)(unsigned long)q->host_address +
+				(index * q->token_size);
 
 	clflush_cache_range(addr, q->token_size);
 
@@ -454,7 +455,7 @@ void *ipu_recv_get_token(struct ipu_fw_com_context *ctx, int q_nbr)
 	if (packets <= 0)
 		return NULL;
 
-	addr = (void *)q->host_address + (rd * q->token_size);
+	addr = (void *)(unsigned long)q->host_address + (rd * q->token_size);
 	clflush_cache_range(addr, q->token_size);
 
 	return addr;
