+++ drivers/media/pci/intel/ipu-isys-queue.c
@@ -671,7 +671,7 @@ static int ipu_isys_stream_start(struct ipu_isys_pipeline *ip,
 					buf, to_dma_addr(msg),
 					sizeof(*buf),
 					IPU_FW_ISYS_SEND_TYPE_STREAM_CAPTURE);
-		ipu_put_fw_mgs_buffer(pipe_av->isys, (u64) buf);
+		ipu_put_fw_mgs_buffer(pipe_av->isys, (uintptr_t) buf);
 	} while (!WARN_ON(rval));
 
 	return 0;
@@ -796,7 +796,7 @@ static void __buf_queue(struct vb2_buffer *vb, bool force)
 				       buf, to_dma_addr(msg),
 				       sizeof(*buf),
 				       IPU_FW_ISYS_SEND_TYPE_STREAM_CAPTURE);
-	ipu_put_fw_mgs_buffer(pipe_av->isys, (u64) buf);
+	ipu_put_fw_mgs_buffer(pipe_av->isys, (uintptr_t) buf);
 	/*
 	 * FIXME: mark the buffers in the buffer list if the queue
 	 * operation fails.
@@ -1374,7 +1374,7 @@ ipu_isys_req_dispatch(struct media_device *mdev,
 				       ip->stream_handle,
 				       set, dma_addr, sizeof(*set),
 				       IPU_FW_ISYS_SEND_TYPE_STREAM_CAPTURE);
-	ipu_put_fw_mgs_buffer(pipe_av->isys, (u64) set);
+	ipu_put_fw_mgs_buffer(pipe_av->isys, (uintptr_t) set);
 
 	WARN_ON(rval);
 }
